+++ sample/src/commonMain/kotlin/com/github/panpf/zoomimage/sample/ui/examples/BasicZoomImageSample.kt
@@ -28,11 +28,11 @@ fun BasicZoomImageSample(sketchImageUri: String, photoPaletteState: MutableState
     BaseZoomImageSample(
         sketchImageUri = sketchImageUri,
         photoPaletteState = photoPaletteState
-    ) { contentScale, alignment, state, scrollBar, onLongClick ->
+    ) { contentScale, alignment, zoomState, scrollBar, onLongClick ->
         val context = LocalPlatformContext.current
         val sketch = SingletonSketch.get(context)
         LaunchedEffect(Unit) {
-            state.subsampling.tileBitmapCache = SketchTileBitmapCache(sketch)
+            zoomState.subsampling.tileBitmapCache = SketchTileBitmapCache(sketch)
         }
 
         var myLoadState by remember { mutableStateOf<MyPageState>(MyPageState.None) }
@@ -52,7 +52,7 @@ fun BasicZoomImageSample(sketchImageUri: String, photoPaletteState: MutableState
                 imageUri = sketchImageUri,
                 http2ByteArray = false
             )
-            state.subsampling.setImageSource(imageSource)
+            zoomState.subsampling.setImageSource(imageSource)
         }
 
         val imagePainter1 = imagePainter
@@ -63,7 +63,7 @@ fun BasicZoomImageSample(sketchImageUri: String, photoPaletteState: MutableState
                 contentScale = contentScale,
                 alignment = alignment,
                 modifier = Modifier.fillMaxSize(),
-                state = state,
+                zoomState = zoomState,
                 scrollBar = scrollBar,
                 onLongPress = {
                     onLongClick.invoke()