+++ sample/src/commonMain/kotlin/com/github/panpf/zoomimage/sample/ui/examples/SketchZoomAsyncImageSample.kt
@@ -39,8 +39,8 @@ fun SketchZoomAsyncImageSample(
             modifier = Modifier
                 .fillMaxSize()
                 .progressIndicator(imageState, rememberSectorProgressPainter()),
-            imageState = imageState,
-            state = state,
+            state = imageState,
+            zoomState = state,
             scrollBar = scrollBar,
             onLongPress = { onLongClick.invoke() }
         )