+++ sample/src/desktopMain/kotlin/com/github/panpf/zoomimage/sample/ui/test/ImageSourceTestScreen.desktop.kt
@@ -13,8 +13,8 @@ import com.github.panpf.sketch.source.ByteArrayDataSource
 import com.github.panpf.sketch.source.FileDataSource
 import com.github.panpf.sketch.util.ioCoroutineDispatcher
 import com.github.panpf.zoomimage.sample.data.ComposeResourceImages
+import com.github.panpf.zoomimage.subsampling.ComposeResourceImageSource
 import com.github.panpf.zoomimage.subsampling.ImageSource
-import com.github.panpf.zoomimage.subsampling.fromComposeResource
 import com.github.panpf.zoomimage.subsampling.fromKotlinResource
 import kotlinx.coroutines.withContext
 import okio.buffer
@@ -63,7 +63,7 @@ actual suspend fun sketchFetcherToZoomImageImageSource(
         }
 
         is ComposeResourceUriFetcher -> {
-            ImageSource.fromComposeResource(fetcher.resourcePath)
+            ComposeResourceImageSource(fetcher.resourcePath)
         }
 
         else -> null