+++ sample/src/jsCommonMain/kotlin/com/github/panpf/zoomimage/sample/ui/test/ImageSourceTestScreen.jsCommon.kt
@@ -10,8 +10,8 @@ import com.github.panpf.sketch.source.ByteArrayDataSource
 import com.github.panpf.sketch.source.FileDataSource
 import com.github.panpf.sketch.util.ioCoroutineDispatcher
 import com.github.panpf.zoomimage.sample.data.ComposeResourceImages
+import com.github.panpf.zoomimage.subsampling.ComposeResourceImageSource
 import com.github.panpf.zoomimage.subsampling.ImageSource
-import com.github.panpf.zoomimage.subsampling.fromComposeResource
 import kotlinx.coroutines.withContext
 import okio.buffer
 import okio.use
@@ -54,7 +54,7 @@ actual suspend fun sketchFetcherToZoomImageImageSource(
         }
 
         is ComposeResourceUriFetcher -> {
-            ImageSource.fromComposeResource(fetcher.resourcePath)
+            ComposeResourceImageSource(fetcher.resourcePath)
         }
 
         else -> null