+++ zoomimage-compose-resources/src/commonMain/kotlin/com/github/panpf/zoomimage/subsampling/ComposeResourceImageSource.kt
@@ -7,23 +7,30 @@ import org.jetbrains.compose.resources.readResourceBytes
 
 /**
  * Create an image source from a compose resource path.
+ *
+ * @param resourcePath The path of the file to read in the compose resource's directory. For example:
+ * * 'composeResources/com.github.panpf.zoomimage.sample.resources/files/huge_china.jpg'
+ * * Res.getUri("files/huge_china.jpg") on android: 'jar:file:/data/app/com.github.panpf.sketch4.sample-1==/base.apk!/composeResources/com.github.panpf.zoomimage.sample.resources/files/huge_china.jpg'
+ * * Res.getUri("files/huge_china.jpg") on desktop: 'file:/Users/panpf/Workspace/zoomimage/sample/build/processedResources/desktop/main/composeResources/com.github.panpf.zoomimage.sample.resources/files/huge_china.jpg'
+ * * Res.getUri("files/huge_china.jpg") on js: 'http://localhost:8080/./composeResources/com.github.panpf.zoomimage.sample.resources/files/huge_china.jpg'
+ * * Res.getUri("files/huge_china.jpg") on ios: 'file:///Users/panpf/Library/Developer/ CoreSimulator/Devices/F828C881-A750-432B-8210-93A84C45E/data/Containers/Bundle/Application/CBD75605-D35E-47A7-B56B-6C5690B062CC/SketchSample.app/compose-resources/composeResources/com.github.panpf.zoomimage.sample.resources/files/huge_china.jpg'
  */
 fun ImageSource.Companion.fromComposeResource(
     resourcePath: String,
 ): ComposeResourceImageSource {
-    return ComposeResourceImageSource(resourcePath)
+    return ComposeResourceImageSource(composeResourceUriToResourcePath(resourcePath))
 }
 
 /**
  * Build a resource path that loads images from the compose resources folder
  *
  * @param resourcePath The path of the file to read in the compose resource's directory. For example:
- * * 'composeResources/com.github.panpf.sketch.sample.resources/files/huge_china.jpg'
- * * Res.getUri("files/huge_china.jpg") on android: 'jar:file:/data/app/com.github.panpf.sketch4.sample-1==/base.apk!/composeResources/com.github.panpf.sketch.sample.resources/files/huge_china.jpg'
- * * Res.getUri("files/huge_china.jpg") on desktop: 'file:/Users/panpf/Workspace/sketch/sample/build/processedResources/desktop/main/composeResources/com.github.panpf.sketch.sample.resources/files/huge_china.jpg'
- * * Res.getUri("files/huge_china.jpg") on js: 'http://localhost:8080/./composeResources/com.github.panpf.sketch.sample.resources/files/huge_china.jpg'
- * * Res.getUri("files/huge_china.jpg") on ios: 'file:///Users/panpf/Library/Developer/ CoreSimulator/Devices/F828C881-A750-432B-8210-93A84C45E/data/Containers/Bundle/Application/CBD75605-D35E-47A7-B56B-6C5690B062CC/SketchSample.app/compose-resources/composeResources/com.github.panpf.sketch.sample.resources/files/huge_china.jpg'
- * @return 'composeResources/com.github.panpf.sketch.sample.resources/files/huge_china.jpg'
+ * * 'composeResources/com.github.panpf.zoomimage.sample.resources/files/huge_china.jpg'
+ * * Res.getUri("files/huge_china.jpg") on android: 'jar:file:/data/app/com.github.panpf.sketch4.sample-1==/base.apk!/composeResources/com.github.panpf.zoomimage.sample.resources/files/huge_china.jpg'
+ * * Res.getUri("files/huge_china.jpg") on desktop: 'file:/Users/panpf/Workspace/zoomimage/sample/build/processedResources/desktop/main/composeResources/com.github.panpf.zoomimage.sample.resources/files/huge_china.jpg'
+ * * Res.getUri("files/huge_china.jpg") on js: 'http://localhost:8080/./composeResources/com.github.panpf.zoomimage.sample.resources/files/huge_china.jpg'
+ * * Res.getUri("files/huge_china.jpg") on ios: 'file:///Users/panpf/Library/Developer/ CoreSimulator/Devices/F828C881-A750-432B-8210-93A84C45E/data/Containers/Bundle/Application/CBD75605-D35E-47A7-B56B-6C5690B062CC/SketchSample.app/compose-resources/composeResources/com.github.panpf.zoomimage.sample.resources/files/huge_china.jpg'
+ * @return 'composeResources/com.github.panpf.zoomimage.sample.resources/files/huge_china.jpg'
  */
 fun composeResourceUriToResourcePath(resourcePath: String): String {
     if (resourcePath.startsWith("composeResources/")) {