+++ zoomimage-compose-sketch3/src/main/kotlin/com/github/panpf/zoomimage/SingletonSketchZoomAsyncImage.kt
@@ -73,7 +73,7 @@ import com.github.panpf.zoomimage.compose.zoom.ScrollBarSpec
  *  rendered onscreen.
  * @param filterQuality Sampling algorithm applied to a bitmap when it is scaled and drawn into the
  *  destination.
- * @param state The state to control zoom
+ * @param zoomState The state to control zoom
  * @param scrollBar Controls whether scroll bars are displayed and their style
  * @param onLongPress Called when the user long presses the image
  * @param onTap Called when the user taps the image
@@ -84,7 +84,7 @@ fun SketchZoomAsyncImage(
     imageUri: String?,
     contentDescription: String?,
     modifier: Modifier = Modifier,
-    imageState: AsyncImageState = rememberAsyncImageState(),
+    state: AsyncImageState = rememberAsyncImageState(),
     placeholder: Painter? = null,
     error: Painter? = null,
     uriEmpty: Painter? = error,
@@ -96,7 +96,7 @@ fun SketchZoomAsyncImage(
     alpha: Float = DefaultAlpha,
     colorFilter: ColorFilter? = null,
     filterQuality: FilterQuality = DrawScope.DefaultFilterQuality,
-    state: ZoomState = rememberZoomState(),
+    zoomState: ZoomState = rememberZoomState(),
     scrollBar: ScrollBarSpec? = ScrollBarSpec.Default,
     onLongPress: ((Offset) -> Unit)? = null,
     onTap: ((Offset) -> Unit)? = null,
@@ -105,7 +105,7 @@ fun SketchZoomAsyncImage(
     contentDescription = contentDescription,
     sketch = LocalContext.current.sketch,
     modifier = modifier,
-    imageState = imageState,
+    state = state,
     placeholder = placeholder,
     error = error,
     uriEmpty = uriEmpty,
@@ -117,7 +117,7 @@ fun SketchZoomAsyncImage(
     alpha = alpha,
     colorFilter = colorFilter,
     filterQuality = filterQuality,
-    state = state,
+    zoomState = zoomState,
     scrollBar = scrollBar,
     onLongPress = onLongPress,
     onTap = onTap,
@@ -160,7 +160,7 @@ fun SketchZoomAsyncImage(
  *  rendered onscreen.
  * @param filterQuality Sampling algorithm applied to a bitmap when it is scaled and drawn into the
  *  destination.
- * @param state The state to control zoom
+ * @param zoomState The state to control zoom
  * @param scrollBar Controls whether scroll bars are displayed and their style
  * @param onLongPress Called when the user long presses the image
  * @param onTap Called when the user taps the image
@@ -171,7 +171,7 @@ fun SketchZoomAsyncImage(
     request: DisplayRequest,
     contentDescription: String?,
     modifier: Modifier = Modifier,
-    imageState: AsyncImageState = rememberAsyncImageState(),
+    state: AsyncImageState = rememberAsyncImageState(),
     placeholder: Painter? = null,
     error: Painter? = null,
     uriEmpty: Painter? = error,
@@ -183,7 +183,7 @@ fun SketchZoomAsyncImage(
     alpha: Float = DefaultAlpha,
     colorFilter: ColorFilter? = null,
     filterQuality: FilterQuality = DrawScope.DefaultFilterQuality,
-    state: ZoomState = rememberZoomState(),
+    zoomState: ZoomState = rememberZoomState(),
     scrollBar: ScrollBarSpec? = ScrollBarSpec.Default,
     onLongPress: ((Offset) -> Unit)? = null,
     onTap: ((Offset) -> Unit)? = null,
@@ -192,7 +192,7 @@ fun SketchZoomAsyncImage(
     contentDescription = contentDescription,
     sketch = LocalContext.current.sketch,
     modifier = modifier,
-    imageState = imageState,
+    state = state,
     placeholder = placeholder,
     error = error,
     uriEmpty = uriEmpty,
@@ -204,7 +204,7 @@ fun SketchZoomAsyncImage(
     alpha = alpha,
     colorFilter = colorFilter,
     filterQuality = filterQuality,
-    state = state,
+    zoomState = zoomState,
     scrollBar = scrollBar,
     onLongPress = onLongPress,
     onTap = onTap,
@@ -241,7 +241,7 @@ fun SketchZoomAsyncImage(
  *  rendered onscreen.
  * @param filterQuality Sampling algorithm applied to a bitmap when it is scaled and drawn into the
  *  destination.
- * @param state The state to control zoom
+ * @param zoomState The state to control zoom
  * @param scrollBar Controls whether scroll bars are displayed and their style
  * @param onLongPress Called when the user long presses the image
  * @param onTap Called when the user taps the image
@@ -252,7 +252,7 @@ fun SketchZoomAsyncImage(
     imageUri: String?,
     contentDescription: String?,
     modifier: Modifier = Modifier,
-    imageState: AsyncImageState = rememberAsyncImageState(),
+    state: AsyncImageState = rememberAsyncImageState(),
     transform: (PainterState) -> PainterState = AsyncImageState.DefaultTransform,
     onPainterState: ((PainterState) -> Unit)? = null,
     alignment: Alignment = Alignment.Center,
@@ -260,7 +260,7 @@ fun SketchZoomAsyncImage(
     alpha: Float = DefaultAlpha,
     colorFilter: ColorFilter? = null,
     filterQuality: FilterQuality = DrawScope.DefaultFilterQuality,
-    state: ZoomState = rememberZoomState(),
+    zoomState: ZoomState = rememberZoomState(),
     scrollBar: ScrollBarSpec? = ScrollBarSpec.Default,
     onLongPress: ((Offset) -> Unit)? = null,
     onTap: ((Offset) -> Unit)? = null,
@@ -269,7 +269,7 @@ fun SketchZoomAsyncImage(
     contentDescription = contentDescription,
     sketch = LocalContext.current.sketch,
     modifier = modifier,
-    imageState = imageState,
+    state = state,
     transform = transform,
     onPainterState = onPainterState,
     alignment = alignment,
@@ -277,7 +277,7 @@ fun SketchZoomAsyncImage(
     alpha = alpha,
     colorFilter = colorFilter,
     filterQuality = filterQuality,
-    state = state,
+    zoomState = zoomState,
     scrollBar = scrollBar,
     onLongPress = onLongPress,
     onTap = onTap,
@@ -318,7 +318,7 @@ fun SketchZoomAsyncImage(
  *  rendered onscreen.
  * @param filterQuality Sampling algorithm applied to a bitmap when it is scaled and drawn into the
  *  destination.
- * @param state The state to control zoom
+ * @param zoomState The state to control zoom
  * @param scrollBar Controls whether scroll bars are displayed and their style
  * @param onLongPress Called when the user long presses the image
  * @param onTap Called when the user taps the image
@@ -329,7 +329,7 @@ fun SketchZoomAsyncImage(
     request: DisplayRequest,
     contentDescription: String?,
     modifier: Modifier = Modifier,
-    imageState: AsyncImageState = rememberAsyncImageState(),
+    state: AsyncImageState = rememberAsyncImageState(),
     transform: (PainterState) -> PainterState = AsyncImageState.DefaultTransform,
     onPainterState: ((PainterState) -> Unit)? = null,
     alignment: Alignment = Alignment.Center,
@@ -337,7 +337,7 @@ fun SketchZoomAsyncImage(
     alpha: Float = DefaultAlpha,
     colorFilter: ColorFilter? = null,
     filterQuality: FilterQuality = DrawScope.DefaultFilterQuality,
-    state: ZoomState = rememberZoomState(),
+    zoomState: ZoomState = rememberZoomState(),
     scrollBar: ScrollBarSpec? = ScrollBarSpec.Default,
     onLongPress: ((Offset) -> Unit)? = null,
     onTap: ((Offset) -> Unit)? = null,
@@ -346,7 +346,7 @@ fun SketchZoomAsyncImage(
     contentDescription = contentDescription,
     sketch = LocalContext.current.sketch,
     modifier = modifier,
-    imageState = imageState,
+    state = state,
     transform = transform,
     onPainterState = onPainterState,
     alignment = alignment,
@@ -354,7 +354,7 @@ fun SketchZoomAsyncImage(
     alpha = alpha,
     colorFilter = colorFilter,
     filterQuality = filterQuality,
-    state = state,
+    zoomState = zoomState,
     scrollBar = scrollBar,
     onLongPress = onLongPress,
     onTap = onTap,